change throughput feature to hosted service on the monitoring instance
